Optimize battery history storage and reduce flash writes
- Change timestamp from uint32_t to uint16_t (saves 2 bytes per entry)
- Skip recording when battery level unchanged and interval < 4x recording interval
- Save history before entering deep sleep mode to prevent data loss
- Track first_record_after_boot to always record on startup
- Add activity state listener for sleep event handling

+++ include/zmk/battery_history/battery_history.h
@@ -13,7 +13,7 @@
  * @brief A single battery history entry
  */
 struct zmk_battery_history_entry {
-    uint32_t timestamp;      // Unix timestamp in seconds
+    uint16_t timestamp;      // Seconds since boot (resets on restart, max ~18 hours)
     uint8_t battery_level;   // Battery percentage (0-100)
 };
